+++ src/services/service_provider.py
@@ -12,7 +12,6 @@ if TYPE_CHECKING:
     from services.clients.cognito import CognitoClient
     from services.clients.event_bridge import EventBridgeClient
     from services.clients.iam import IAMClient
-    from services.clients.modular import ModularClient
     from services.clients.s3 import ModularAssumeRoleS3Service
     from services.clients.s3 import S3Client
     from services.clients.scheduler import EventBridgeJobScheduler
@@ -43,6 +42,7 @@ if TYPE_CHECKING:
     from services.chronicle_service import ChronicleInstanceService
     from services.reports import ReportMetricsService
     from services.metadata import MetadataProvider
+    from modular_sdk.modular import ModularServiceProvider
 
 
 class ServiceProvider(metaclass=SingletonMeta):
@@ -57,10 +57,7 @@ class ServiceProvider(metaclass=SingletonMeta):
     @cached_property
     def s3(self) -> 'S3Client':
         from services.clients.s3 import S3Client
-        env = self.environment_service
-        if env.is_docker():
-            return S3Client.factory().build_minio()
-        return S3Client.factory().build_s3(env.aws_region())
+        return S3Client.build()
 
     @cached_property
     def ssm(self) -> 'CachedSSMClient':
@@ -77,9 +74,7 @@ class ServiceProvider(metaclass=SingletonMeta):
         from services.clients.sts import StsClient
         if self.environment_service.is_docker():
             return StsClient.build()
-        return StsClient.factory().build(
-            region_name=self.environment_service.aws_region()
-        )
+        return StsClient.build()
 
     @cached_property
     def batch(self) -> Union['BatchClient', 'CeleryJobClient']:
@@ -87,9 +82,7 @@ class ServiceProvider(metaclass=SingletonMeta):
             from services.clients.batch import CeleryJobClient
             return CeleryJobClient.build()
         from services.clients.batch import BatchClient
-        return BatchClient.factory().build(
-            region_name=self.environment_service.aws_region()
-        )
+        return BatchClient.build()
 
     @cached_property
     def onprem_users_client(self) -> 'MongoAndSSMAuthClient':
@@ -113,16 +106,14 @@ class ServiceProvider(metaclass=SingletonMeta):
         return LambdaClient(environment_service=self.environment_service)
 
     @cached_property
-    def modular_client(self) -> 'ModularClient':
-        from services.clients.modular import ModularClient
-        return ModularClient()
+    def modular_client(self) -> 'ModularServiceProvider':
+        from modular_sdk.modular import ModularServiceProvider
+        return ModularServiceProvider()
 
     @cached_property
     def events(self) -> 'EventBridgeClient':
         from services.clients.event_bridge import EventBridgeClient
-        return EventBridgeClient.factory().build(
-            region_name=self.environment_service.aws_region()
-        )
+        return EventBridgeClient.build()
 
     @cached_property
     def iam(self) -> 'IAMClient':